requirements.txt
txt

# requirements.txt

 

# Основные зависимости

mysql-connector-python==8.0.33

SQLAlchemy==2.0.19

python-dotenv==1.0.0

 

# Дополнительные зависимости вашего приложения

requests==2.31.0

pandas==2.1.3

numpy==1.24.3

click==8.1.7

 

# Веб-фреймворк (если используется)

Flask==3.0.0

# или

FastAPI==0.104.1

uvicorn[standard]==0.24.0
